# docs/swagger.yaml
info:
  contact: {}
paths:
  /auth/delete:
    delete:
      description: Deletes the account and signs out the user
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Delete the account
      tags:
      - Auth
  /auth/signin:
    post:
      consumes:
      - application/json
      description: Signs in the user, returns the user details and sets the jwt token
      parameters:
      - description: Login details
        in: body
        name: user
        required: true
        schema:
          properties:
            email:
              type: string
            password:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              user:
                properties:
                  created_at:
                    type: string
                  email:
                    type: string
                  first_name:
                    type: string
                  last_name:
                    type: string
                  role:
                    type: string
                  user_id:
                    type: string
                type: object
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Sign in
      tags:
      - Auth
...
  /auth/signup:
    post:
      consumes:
      - application/json
      description: Create a new user
      parameters:
      - description: User details
        in: body
        name: user
        required: true
        schema:
          properties:
            email:
              type: string
            first_name:
              type: string
            last_name:
              type: string
            password:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            properties:
              user:
                properties:
                  created_at:
                    type: string
                  email:
                    type: string
                  first_name:
                    type: string
                  last_name:
                    type: string
                  role:
                    type: string
                  user_id:
                    type: string
                type: object
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "409":
          description: Conflict
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Create a new user
      tags:
      - Auth
  /cart/create:
    post:
      consumes:
      - application/json
      description: Create a new cart item
      parameters:
      - description: Cart item details
        in: body
        name: cart_item
        required: true
        schema:
          properties:
            price:
              type: number
            product_id:
              type: integer
            quantity:
              type: integer
            user_id:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            properties:
              cartItem:
                properties:
                  created_at:
                    type: string
                  id:
                    type: integer
                  price:
                    type: number
                  product_id:
                    type: integer
                  quantity:
                    type: integer
                  user_id:
                    type: string
                type: object
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Create a new cart item
      tags:
      - Cart
  /cart/delete/{id}:
    delete:
      description: Delete a cart item
      parameters:
      - description: Cart Item ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Delete a cart item
      tags:
      - Cart
  /cart/empty/{id}:
    delete:
      description: Empties the cart of a user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Empty the cart
      tags:
      - Cart
  /cart/list:
    get:
      description: Get all cart items
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              cartItems:
                items:
                  properties:
                    created_at:
                      type: string
                    id:
                      type: integer
                    price:
                      type: number
                    product_id:
                      type: integer
                    quantity:
                      type: integer
                    user_id:
                      type: string
                  type: object
                type: array
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Get all cart items
      tags:
      - Cart
  /cart/show-by-user/{id}:
    get:
      description: Get all cart items by user ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              cartItems:
                items:
                  properties:
                    created_at:
                      type: string
                    id:
                      type: integer
                    price:
                      type: number
                    product_id:
                      type: integer
                    quantity:
                      type: integer
                    user_id:
                      type: string
                  type: object
                type: array
            type: object
        "404":
          description: Not Found
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Get all cart items by user ID
      tags:
      - Cart
  /cart/update/{id}:
    put:
      consumes:
      - application/json
      description: Update a cart item
      parameters:
      - description: Cart Item ID
        in: path
        name: id
        required: true
        type: integer
      - description: Cart item details
        in: body
        name: cart_item
        required: true
        schema:
          properties:
            price:
              type: number
            product_id:
              type: integer
            quantity:
              type: integer
            user_id:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              cartItem:
                properties:
                  created_at:
                    type: string
                  id:
                    type: integer
                  price:
                    type: number
                  product_id:
                    type: integer
                  quantity:
                    type: integer
                  user_id:
                    type: string
                type: object
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Update a cart item
      tags:
      - Cart
...
swagger: "2.0"